Migrate LogViewer off redux.
