+++ src/components/LogViewer/types.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,4 +14,21 @@
  * limitations under the License.
  */
 
-export const placeholder = true;
+export interface LogEntry {
+  level: string;
+  message: string;
+  timestamp: number;
+  data: {
+    user: { [key: string]: any };
+    metadata?: {
+      level: string;
+      message: string;
+      function?: {
+        name: string;
+      };
+      emulator?: {
+        name: string;
+      };
+    };
+  };
+}
