--- src/store/logviewer/reducer.ts
@@ -1,30 +1,0 @@
-/**
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { Action, createReducer } from 'typesafe-actions';
-
-import { logReceived, logReset } from './actions';
-import { LogState } from './types';
-
-export const logReducer = createReducer<LogState, Action>({
-  history: [],
-})
-  .handleAction(logReceived, (state: LogState, { payload }) => {
-    return { ...state, history: [...state.history, payload] };
-  })
-  .handleAction(logReset, (state: LogState) => {
-    return { ...state, history: [] };
-  });
