--- src/store/logviewer/types.ts
@@ -1,38 +1,0 @@
-/**
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export interface LogState {
-  history: LogEntry[];
-}
-
-export interface LogEntry {
-  level: string;
-  message: string;
-  timestamp: number;
-  data: {
-    user: { [key: string]: any };
-    metadata?: {
-      level: string;
-      message: string;
-      function?: {
-        name: string;
-      };
-      emulator?: {
-        name: string;
-      };
-    };
-  };
-}
